updates to module format
