+++ .eslintignore
@@ -1,1 +1,3 @@
 node_modules/
+dist/
+__test__/
