+++ cspell.json
@@ -6,6 +6,8 @@
     "touchpoint",
     "TOUCHPOINT",
     "jmassardo",
-    "totango"
+    "totango",
+    "codespace",
+    "tada"
   ]
 }
