+++ package.json
@@ -4,7 +4,8 @@
   "description": "This action creates a Totango touchpoint when a new issue is created",
   "main": "index.js",
   "scripts": {
-    "test": "eslint --ignore-path .eslintignore . ; cspell *.js *.md"
+    "test": "eslint --ignore-path .eslintignore . ; cspell *.js *.md",
+    "build": "ncc build src/index.js --license licenses.txt"
   },
   "repository": {
     "type": "git",
@@ -20,14 +21,18 @@
   "dependencies": {
     "@actions/core": "^1.10.0",
     "@actions/github": "^5.1.1",
-    "cspell": "^6.17.0",
-    "eslint": "^8.30.0",
-    "eslint-config-strongloop": "^2.1.0",
     "request": "^2.88.2",
     "showdown": "^2.1.0",
-    "uuid": "^7.0.0"
+    "uuid": "^9.0.0"
   },
   "directories": {
     "example": "examples"
+  },
+  "devDependencies": {
+    "chai": "^4.3.7",
+    "cspell": "^6.19.2",
+    "eslint": "^8.33.0",
+    "eslint-config-strongloop": "^2.1.0",
+    "mocha": "^10.2.0"
   }
 }
