+++ src/github.js
@@ -1,0 +1,7 @@
+const core = require('@actions/core');
+const github = require('@actions/github');
+
+module.exports = {
+  core,
+  github,
+};
