+++ utils/build.sh
@@ -2,5 +2,5 @@
 
 npm install @actions/core
 npm install @actions/github
-ncc build index.js --license licenses.txt
+ncc build src/index.js --license licenses.txt
 rm -rf node_modules
